backend/requirements.txt
fastapi[standard]
passlib[bcrypt]
pip
psycopg[binary]
pyjwt
sqlmodel
